// types/tabs.ts
// export type Tab = {
//   id: string;
//   title: string;
//   url: string;
//   icon: string;
//   isActive: boolean;
//   isPinned: boolean;
//   isPrivate: boolean;
//   isLoading: boolean;
//   isSuspended: boolean;
// };

export type Tab = {
  id: string;
  image: string;
  color: string;
};